requirements.txt
click
matplotlib
py-trello
pyyaml
